// scrape/sofa/types.ts
export interface Incident {
    text?:               string;
    homeScore?:          number;
    awayScore?:          number;
    isLive?:             boolean;
    time:                number;
    addedTime?:          number;
    incidentType:        string;
    playerIn?:           Player;
    playerOut?:          Player;
    id?:                 number;
    isHome?:             boolean;
    reversedPeriodTime?: number;
    length?:             number;
    player?:             Player;
    playerName?:         string;
    reason?:             string;
    rescinded?:          boolean;
    incidentClass?:      string;
    injury?:             boolean;
    assist1?:            Player;
}

export interface Player {
    name:       string;
    firstName?: string;
    lastName?:  string;
    slug:       string;
    shortName:  string;
    position:   Position;
    userCount:  number;
    id:         number;
}


export enum Position {
    D = "D",
    F = "F",
    M = "M",
}